test: split scaffold.test.js into per-type files with shared makeTmpDir helper
